Add guild fields to save functions for proper persistence
- db-manager.js: Add guild_id, guild_contribution, guild_last_active to rpcData
- event-save-manager.js: Add same guild fields to playerData
- Fix type casting in migration 021: guild_id → INTEGER, guild_contribution → BIGINT

Without these changes, guild membership was not being saved to the database
because the fields were missing from the RPC call payload.

+++ database/migrations/021_fix_guild_fields_in_rpc.sql
@@ -108,12 +108,12 @@ BEGIN
         -- GUILD FIELDS (previously missing!)
         guild_id = CASE
             WHEN p_data ? 'guild_id'
-            THEN (p_data->>'guild_id')::BIGINT
+            THEN (p_data->>'guild_id')::INTEGER
             ELSE guild_id
         END,
         guild_contribution = CASE
             WHEN p_data ? 'guild_contribution'
-            THEN COALESCE((p_data->>'guild_contribution')::INTEGER, guild_contribution)
+            THEN COALESCE((p_data->>'guild_contribution')::BIGINT, guild_contribution)
             ELSE guild_contribution
         END,
         guild_last_active = CASE
